하나의 Controller 에서  Comment 와 Reply 두개로 분리 refactoring
