+++ backend/src/main/java/com/example/Main/domain/Post/Comment/controller/ApiV1PostReplyController.java
@@ -1,14 +1,8 @@
 package com.example.Main.domain.Post.Comment.controller;
 
-import com.example.Main.domain.Member.entity.Member;
-import com.example.Main.domain.Member.service.MemberService;
 import com.example.Main.domain.Post.Comment.dto.PostCommentDTO;
 import com.example.Main.domain.Post.Comment.dto.request.PostCommentCreateRequest;
 import com.example.Main.domain.Post.Comment.dto.request.PostCommentModifyRequest;
-import com.example.Main.domain.Post.Comment.dto.response.CommentCreateResponse;
-import com.example.Main.domain.Post.Comment.dto.response.CommentModifyResponse;
-import com.example.Main.domain.Post.Comment.dto.response.CommentResponse;
-import com.example.Main.domain.Post.Comment.dto.response.CommentsResponse;
 import com.example.Main.domain.Post.Comment.entity.PostComment;
 import com.example.Main.domain.Post.Comment.service.PostCommentService;
 import com.example.Main.domain.Post.entity.Post;
@@ -24,187 +18,14 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(value = "/api/v1/post/{postId}/comments")
-public class ApiV1CommentPostController {
+@RequestMapping(value = "/api/v1/post/{postId}/comments/{commentId}/replies")
+public class ApiV1PostReplyController {
 
     private final PostCommentService commentService;
-    private final MemberService memberService;
     private final PostService postService;
 
-    // 특정 게시글의 댓글 목록 조회 (다건 조회)
-    @GetMapping
-    public RsData<CommentsResponse> getComments(@PathVariable("postId") Long postId) {
-        Post post = postService.getPost(postId);
-        if (post == null) {
-            return RsData.of("404", "게시글을 찾을 수 없습니다.", null);
-        }
-
-        List<PostCommentDTO> comments = commentService.getCommentsByPostId(postId);
-        if (comments.isEmpty()) {
-            return RsData.of("404", "댓글이 없습니다.", null);
-        }
-
-        return RsData.of("200", "댓글 조회 성공 (게시글 제목: " + post.getSubject() + ")", new CommentsResponse(comments));
-    }
-
-    // 특정 게시글의 댓글 조회 (단건 조회)
-    @GetMapping("/{commentId}")
-    public RsData<PostCommentDTO> getComment(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId) {
-        Post post = postService.getPost(postId);
-        if (post == null) {
-            return RsData.of("404", "게시글을 찾을 수 없습니다.", null);
-        }
-
-        PostComment comment = commentService.getComment(commentId).orElse(null);
-        if (comment == null || !comment.getPost().getId().equals(postId)) {
-            return RsData.of("404", "댓글이 존재하지 않거나, 게시글과 일치하지 않습니다.", null);
-        }
-
-        return RsData.of("200", "댓글 조회 성공 (게시글 제목: " + post.getSubject() + ")", new PostCommentDTO(comment));
-    }
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/mycomments")
-    public RsData<CommentsResponse> getMyPostComments(@PathVariable("postId") Long postId, Principal principal) {
-        if (principal == null) {
-            return RsData.of("401", "로그인 후 사용 가능합니다.", null);
-        }
-
-        String loggedInUserEmail = principal.getName();
-        Post post = postService.getPost(postId);
-        if (post == null) {
-            return RsData.of("404", "게시글이 존재하지 않습니다.", null);
-        }
-
-        List<PostCommentDTO> myPostComments = commentService.getPostsCommentsByUserAndPostId(loggedInUserEmail, postId);
-        if (myPostComments.isEmpty()) {
-            return RsData.of("404", "본인이 작성한 댓글이 없습니다.", null);
-        }
-
-        return RsData.of("200", "본인이 작성한 댓글 조회 성공", new CommentsResponse(myPostComments));
-    }
-
-    // 게시글 댓글 작성
-    @PreAuthorize("isAuthenticated()")
-    @PostMapping
-    public RsData<CommentCreateResponse> postCommentCreate(@PathVariable("postId") Long postId,
-                                                           @Valid @RequestBody PostCommentCreateRequest commentCreateRequest,
-                                                           Principal principal) {
-        if (principal == null) {
-            return RsData.of("401", "로그인 후 사용 가능합니다.", null);
-        }
-
-        String userEmail = principal.getName();
-        Long parentCommentId = commentCreateRequest.getParentCommentId();
-
-        PostComment comment = commentService.addComment(postId, userEmail, commentCreateRequest.getContent(), parentCommentId);
-
-        if (comment == null) {
-            return RsData.of("404", "게시글이 존재하지 않습니다.", null);
-        }
-
-        // CommentCreateResponse 사용
-        CommentCreateResponse response = new CommentCreateResponse(comment);
-        return RsData.of("201", "댓글 작성 성공", response);
-    }
-
-
-    // 게시글 댓글 수정
-    @PreAuthorize("isAuthenticated()")
-    @PatchMapping("/{commentId}")
-    public RsData<CommentModifyResponse> postCommentModify(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId, @Valid @RequestBody PostCommentModifyRequest commentModifyRequest, Principal principal) {
-        if (principal == null) {
-            return RsData.of("401", "로그인 후 사용 가능합니다.", null);
-        }
-
-        String userEmail = principal.getName();
-        PostComment comment = commentService.getComment(commentId).orElse(null);
-
-        if (comment == null) {
-            return RsData.of("404", "댓글을 찾을 수 없습니다.", null);
-        }
-
-        if (!comment.getPost().getId().equals(postId)) {
-            return RsData.of("404", "댓글이 속한 게시글 번호가 일치하지 않습니다.", null);
-        }
-
-
-        if (!comment.getAuthor().getEmail().equals(userEmail)) {
-            return RsData.of("403", "본인만 댓글을 수정할 수 있습니다.", null);
-        }
-
-
-        comment = commentService.updateComment(commentId, commentModifyRequest.getContent(), userEmail);
-
-        return RsData.of("200", "댓글 수정 성공", new CommentModifyResponse(comment));
-    }
-
-    // 게시글 댓글 삭제
-    @PreAuthorize("isAuthenticated()")
-    @DeleteMapping("/{commentId}")
-    public RsData<String> deleteComment(@PathVariable("postId") Long postId,
-                                        @PathVariable("commentId") Long commentId,
-                                        Principal principal) {
-        if (principal == null) {
-            return RsData.of("401", "로그인 후 사용 가능합니다.", null);
-        }
-
-        String loggedInUser = principal.getName();
-        PostComment comment = commentService.getComment(commentId).orElse(null);
-
-        if (comment == null) {
-            return RsData.of("404", "댓글이 존재하지 않습니다.", null);
-        }
-
-        if (!comment.getPost().getId().equals(postId)) {
-            return RsData.of("404", "댓글이 속한 게시글 번호가 일치하지 않습니다.", null);
-        }
-
-        if (!comment.getAuthor().getEmail().equals(loggedInUser)) {
-            return RsData.of("403", "본인만 댓글을 삭제할 수 있습니다.", null);
-        }
-
-        if (commentService.hasReplies(comment)) {
-            return RsData.of("400", "대댓글이 달린 댓글은 삭제할 수 없습니다.", null);
-        }
-
-        commentService.deleteComment(commentId);
-        return RsData.of("200", "%d 번 댓글 삭제 성공".formatted(commentId), null);
-    }
-
-    // 게시글 댓글 좋아요
-    @PostMapping("/{commentId}/like")
-    public RsData<CommentResponse> like(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId,
-                                        Principal principal) {
-        if (principal == null) {
-            return RsData.of("401", "로그인 후 사용 가능합니다.", null);
-        }
-
-        String loggedInUser = principal.getName();
-        PostComment comment = commentService.getComment(commentId).orElse(null);
-
-        if (comment == null) {
-            return RsData.of("404", "댓글이 존재하지 않거나 임시 저장된 댓글입니다.", null);
-        }
-
-        if (!comment.getPost().getId().equals(postId)) {
-            return RsData.of("404", "댓글이 해당 게시글과 일치하지 않습니다.", null);
-        }
-
-        Member member = memberService.getMemberByEmail(loggedInUser);
-        boolean isLiked = comment.getLikedByMembers().contains(member);
-
-        if (isLiked) {
-            commentService.unlikeComment(commentId, loggedInUser);
-            return RsData.of("200", "댓글 좋아요 취소", new CommentResponse(new PostCommentDTO(comment)));
-        } else {
-            commentService.likeComment(commentId, loggedInUser);
-            return RsData.of("200", "댓글 좋아요 성공", new CommentResponse(new PostCommentDTO(comment)));
-        }
-    }
-
     // 대댓글 조회
-    @GetMapping("/{commentId}/replies")
+    @GetMapping
     public RsData<List<PostCommentDTO>> getReplies(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId) {
 
         Post post = postService.getPost(postId);
@@ -223,7 +44,7 @@ public class ApiV1CommentPostController {
     }
 
     // 대댓글 단건 조회
-    @GetMapping("/{commentId}/replies/{replyId}")
+    @GetMapping("/{replyId}")
     public RsData<PostCommentDTO> getReply(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId,
                                            @PathVariable("replyId") Long replyId) {
 
@@ -255,7 +76,7 @@ public class ApiV1CommentPostController {
 
     // 본인이 작성한 대댓글 조회
     @PreAuthorize("isAuthenticated()")
-    @GetMapping("/{commentId}/myreplies")
+    @GetMapping("/myreplies")
     public RsData<List<PostCommentDTO>> getMyReplies(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId, Principal principal) {
         if (principal == null) {
             return RsData.of("401", "로그인 후 사용 가능합니다.", null);
@@ -290,7 +111,7 @@ public class ApiV1CommentPostController {
 
     // 대댓글 작성
     @PreAuthorize("isAuthenticated()")
-    @PostMapping("/{commentId}/replies")
+    @PostMapping
     public RsData<PostCommentDTO> postReplyCreate(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId,
                                                   @Valid @RequestBody PostCommentCreateRequest commentCreateRequest,
                                                   Principal principal) {
@@ -331,7 +152,7 @@ public class ApiV1CommentPostController {
 
     // 대댓글 수정
     @PreAuthorize("isAuthenticated()")
-    @PatchMapping("/{commentId}/replies/{replyId}")
+    @PatchMapping("/{replyId}")
     public RsData<PostCommentDTO> modifyReply(@PathVariable("postId") Long postId, @PathVariable("commentId") Long commentId,
                                               @PathVariable("replyId") Long replyId,
                                               @Valid @RequestBody PostCommentModifyRequest commentModifyRequest,
@@ -375,7 +196,7 @@ public class ApiV1CommentPostController {
 
     // 대댓글 삭제
     @PreAuthorize("isAuthenticated()")
-    @DeleteMapping("/{commentId}/replies/{replyId}")
+    @DeleteMapping("/{replyId}")
     public RsData<String> deleteReply(@PathVariable("postId") Long postId,
                                       @PathVariable("commentId") Long commentId,
                                       @PathVariable("replyId") Long replyId,
