+++ backend/src/main/java/com/example/Main/domain/Post/Comment/dto/response/PostCommentResponse.java
@@ -4,10 +4,9 @@ import com.example.Main.domain.Post.Comment.dto.PostCommentDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.util.List;
-
 @Getter
 @AllArgsConstructor
-public class CommentsResponse {
-    private final List<PostCommentDTO> commentDTOS;
+public class PostCommentResponse {
+    private final PostCommentDTO comment;
 }
+
