+++ backend/src/main/java/com/example/Main/domain/QnA/Comment/dto/response/QnACommentsResponse.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class CommentsResponse {
+public class QnACommentsResponse {
     private final List<QnACommentDTO> commentDTOS;
 }
